backend/backend/requirements.txt
bs4
django
djnagorestframework
django-cors-headsers
requests
lxml
mongoengine
django-rest-framework-mongoengine
nest-asyncio
asyncio
geopy
celery
channels
channels_redis
django-rest-knox